+++ Modules/Intent.Modules.Common.CSharp/FactoryExtensions/KnownCSharpTypesCache.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using Intent.Engine;
 using Intent.Modules.Common.Plugins;
 using Intent.Modules.Common.Templates;
@@ -26,10 +27,33 @@ namespace Intent.Modules.Common.CSharp.FactoryExtensions
         /// <inheritdoc />
         protected override void OnAfterTemplateRegistrations(IApplication application)
         {
-            _knownTypesByNamespace = application
+            var knownTypesByNamespace = application
                 .FindTemplateInstances<IClassProvider>(TemplateDependency.OfType<IClassProvider>())
                 .GroupBy(x => string.IsNullOrWhiteSpace(x.Namespace) ? string.Empty : x.Namespace)
                 .ToDictionary(x => x.Key, x => (ISet<string>)new HashSet<string>(x.Select(y => y.ClassName)));
+            AddCommonKnownTypes(knownTypesByNamespace);
+            _knownTypesByNamespace = knownTypesByNamespace;
+        }
+
+        private void AddCommonKnownTypes(Dictionary<string, ISet<string>> knownTypesByNamespace)
+        {
+            AddCommonKnownType(knownTypesByNamespace, "System.Attribute");
+            AddCommonKnownType(knownTypesByNamespace, "System.Action");
+        }
+
+        private void AddCommonKnownType(Dictionary<string, ISet<string>> knownTypesByNamespace, string fullTypeName)
+        {
+            int classStartIndex = fullTypeName.LastIndexOf(".");
+            var namespaceName = fullTypeName.Substring(0, classStartIndex);
+            var className = fullTypeName.Substring(classStartIndex + 1); 
+
+            if (!knownTypesByNamespace.TryGetValue(namespaceName, out var systemKnownTypes))
+            {
+                systemKnownTypes = new HashSet<string>();
+                knownTypesByNamespace.Add(namespaceName, systemKnownTypes);
+            }
+
+            systemKnownTypes.Add(className);
         }
 
         /// <summary>