+++ Modules/Intent.Modules.Modelers.Services.GraphQL/Api/GraphQLSubscriptionModel.cs
@@ -1,0 +1,85 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Intent.Metadata.Models;
+using Intent.Modules.Common;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.Api.ApiElementModel", Version = "1.0")]
+
+namespace Intent.Modelers.Services.GraphQL.Api
+{
+    [IntentManaged(Mode.Fully, Signature = Mode.Fully)]
+    public class GraphQLSubscriptionModel : IMetadataModel, IHasStereotypes, IHasName, IHasTypeReference
+    {
+        public const string SpecializationType = "GraphQL Subscription";
+        public const string SpecializationTypeId = "37e9cf40-ac8e-4880-9a4a-92b540a4fea7";
+        protected readonly IElement _element;
+
+        [IntentManaged(Mode.Fully)]
+        public GraphQLSubscriptionModel(IElement element, string requiredType = SpecializationType)
+        {
+            if (!requiredType.Equals(element.SpecializationType, StringComparison.InvariantCultureIgnoreCase))
+            {
+                throw new Exception($"Cannot create a '{GetType().Name}' from element with specialization type '{element.SpecializationType}'. Must be of type '{SpecializationType}'");
+            }
+            _element = element;
+        }
+
+        public string Id => _element.Id;
+
+        public string Name => _element.Name;
+
+        public string Comment => _element.Comment;
+
+        public IEnumerable<IStereotype> Stereotypes => _element.Stereotypes;
+
+        public ITypeReference TypeReference => _element.TypeReference;
+
+        public IElement InternalElement => _element;
+
+        public GraphQLEventMessageModel EventMessage => _element.ChildElements
+            .GetElementsOfType(GraphQLEventMessageModel.SpecializationTypeId)
+            .Select(x => new GraphQLEventMessageModel(x))
+            .SingleOrDefault();
+
+        public override string ToString()
+        {
+            return _element.ToString();
+        }
+
+        public bool Equals(GraphQLSubscriptionModel other)
+        {
+            return Equals(_element, other?._element);
+        }
+
+        public override bool Equals(object obj)
+        {
+            if (ReferenceEquals(null, obj)) return false;
+            if (ReferenceEquals(this, obj)) return true;
+            if (obj.GetType() != this.GetType()) return false;
+            return Equals((GraphQLSubscriptionModel)obj);
+        }
+
+        public override int GetHashCode()
+        {
+            return (_element != null ? _element.GetHashCode() : 0);
+        }
+    }
+
+    [IntentManaged(Mode.Fully)]
+    public static class GraphQLSubscriptionModelExtensions
+    {
+
+        public static bool IsGraphQLSubscriptionModel(this ICanBeReferencedType type)
+        {
+            return type != null && type is IElement element && element.SpecializationTypeId == GraphQLSubscriptionModel.SpecializationTypeId;
+        }
+
+        public static GraphQLSubscriptionModel AsGraphQLSubscriptionModel(this ICanBeReferencedType type)
+        {
+            return type.IsGraphQLSubscriptionModel() ? new GraphQLSubscriptionModel((IElement)type) : null;
+        }
+    }
+}