+++ Modules/Intent.Modules.Common.CSharp/Builder/CSharpConstructor.cs
@@ -4,13 +4,16 @@ using System.Linq;
 
 namespace Intent.Modules.Common.CSharp.Builder;
 
-public class CSharpConstructor : CSharpMember<CSharpConstructor>
+public class CSharpConstructor : CSharpMember<CSharpConstructor>, IHasCSharpStatements
 {
     public CSharpClass Class { get; }
     public string AccessModifier { get; private set; } = "public ";
     public CSharpConstructorCall ConstructorCall { get; private set; }
     public IList<CSharpConstructorParameter> Parameters { get; } = new List<CSharpConstructorParameter>();
     public List<CSharpStatement> Statements { get; } = new();
+
+    IList<CSharpStatement> IHasCSharpStatements.Statements => this.Statements;
+
     public CSharpConstructor(CSharpClass @class)
     {
         BeforeSeparator = CSharpCodeSeparatorType.EmptyLines;