+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,5 +1,6 @@
 ﻿### Version 3.3.34
 
+- Update: `.Async()` on Class Methods now checks if return type is of type `Task` and adds it if not specified.
 - Fixed: Comments for interfaces themselves would not be generated.
 
 ### Version 3.3.33