+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,6 +1,9 @@
 ﻿### Version 3.3.34
 
 - Fixed: Comments for interfaces themselves would not be generated.
+- New: Support for .Net known type to improve `Type` name normalization, currently supports `System.Attribute`, `System.Action`.
+- Fixed: `Type` name normalization system was unaware of `CSharpFileBuilder` using clauses.
+- Fixed: `Type` name normalization wasn't disambiguating `Type` names and `Namespaces` under certain conditions.
 
 ### Version 3.3.33