+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -4,6 +4,7 @@
 - New: Support for .Net known type to improve `Type` name normalization, currently supports `System.Attribute`, `System.Action`.
 - Fixed: `Type` name normalization system was unaware of `CSharpFileBuilder` using clauses.
 - Fixed: `Type` name normalization wasn't disambiguating `Type` names and `Namespaces` under certain conditions.
+- Fixed: Under certain circumstances `CSharpFileBuilder`'s setters and getters would generate the requested access modifiers.
 
 ### Version 3.3.33